% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataUnit.R
\docType{data}
\name{dataUnit}
\alias{dataUnit}
\title{Data Unit Level of Rice Field Productivity}
\format{
A data frame with 210 observations on the following 7 variables:
\describe{
  \item{number}{order of observation}
  \item{area}{domain codes}
  \item{name}{name of the small area}
  \item{y}{rice field productivity}
  \item{x1}{proportion of paddy rice harvested area to total harvested area}
  \item{x2}{latitude}
  \item{x3}{longitude}
}
}
\usage{
dataUnit
}
\description{
This dataset is data on rice productivity in 15 sub-districts from 3 districts (Seruyan, East Kotawaringin, and West Kotawaringin) in Central Kalimantan obtained from the Ubinan Survey conducted routinely by BPS. This data will be implemented with the Geoadditive Small Area Model
}
\keyword{datasets}
