% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eblupGeo.R
\name{eblupgeo}
\alias{eblupgeo}
\title{EBLUP's for domain means using Geoadditive Small Area Model}
\usage{
eblupgeo(formula, zspline, dom, xmean, zmean, data)
}
\arguments{
\item{formula}{the model that to be fitted}

\item{zspline}{matrix that used in model for random effect of spline-2}

\item{dom}{domain codes}

\item{xmean}{matrix of auxiliary variables means for each domains}

\item{zmean}{matrix of spline-2 means for each domains}

\item{data}{data unit level that used as data frame that containing the variables named in formula and dom}
}
\description{
This function calculates EBLUP's based on unit level using Geoadditive Small Area Model
}
