% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbmseGeo.R
\name{pbmsegeo}
\alias{pbmsegeo}
\title{Parametric Bootstrap Mean Squared Error of EBLUP's for domain means using Geoadditive Small Area Model}
\usage{
pbmsegeo(formula, zspline, dom, xmean, zmean, data, B = 100)
}
\arguments{
\item{formula}{the model that to be fitted}

\item{zspline}{matrix that used in model for random effect of spline-2}

\item{dom}{domain codes}

\item{xmean}{matrix of auxiliary variables means for each domains}

\item{zmean}{matrix of spline-2 means for each domains}

\item{data}{data unit level that used as data frame that containing the variables named in formula and dom}

\item{B}{iteration bootstraping}
}
\description{
This function calculates MSE of EBLUP's based on unit level using Geoadditive Small Area Model
}
