% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zspline.R
\docType{data}
\name{zspline}
\alias{zspline}
\title{Z-Spline}
\format{
A data frame with 210 observations on the following 6 variables (number of knots used)
}
\usage{
zspline
}
\description{
This dataset is obtained from the calculation of the optimum GCV (Generalized Cross Validation), where there are 6 knots that have the lowest GCV value.
}
\keyword{datasets}
